+++ Block/Picture.php
@@ -18,9 +18,14 @@ class Picture extends Template
     private $images = [];
 
     /**
+     * @var Image
+     */
+    private $originalImage;
+
+    /**
      * @var string
      */
-    private $originalImage = '';
+    private $title = '';
 
     /**
      * @var string
@@ -36,6 +41,11 @@ class Picture extends Template
      * @var string
      */
     private $height = '';
+
+    /**
+     * @var string
+     */
+    private $style = '';
 
     /**
      * @var string
@@ -71,7 +81,7 @@ class Picture extends Template
     }
 
     /**
-     * @param array $images
+     * @param Image[] $images
      * @return Picture
      */
     public function setImages(array $images): Picture
@@ -91,21 +101,40 @@ class Picture extends Template
     }
 
     /**
-     * @return string
+     * @return Image
      */
-    public function getOriginalImage(): string
+    public function getOriginalImage(): Image
     {
         return $this->originalImage;
     }
 
     /**
-     * @param string $originalImage
+     * @param Image $originalImage
      *
      * @return Picture
      */
-    public function setOriginalImage(string $originalImage): Picture
+    public function setOriginalImage(Image $originalImage): Picture
     {
         $this->originalImage = $originalImage;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getTitle(): string
+    {
+        return $this->title;
+    }
+
+    /**
+     * @param string $title
+     *
+     * @return Picture
+     */
+    public function setTitle(string $title): Picture
+    {
+        $this->title = $title;
         return $this;
     }
 
@@ -167,6 +196,24 @@ class Picture extends Template
     /**
      * @return string
      */
+    public function getStyle(): string
+    {
+        return $this->style;
+    }
+
+    /**
+     * @param string $style
+     * @return Picture
+     */
+    public function setStyle(string $style): Picture
+    {
+        $this->style = $style;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
     public function getOriginalTag(): string
     {
         return $this->originalTag;
@@ -188,11 +235,11 @@ class Picture extends Template
      */
     public function getOriginalImageType(): string
     {
-        if (preg_match('/\.(jpg|jpeg)$/i', $this->getOriginalImage())) {
+        if (preg_match('/\.(jpg|jpeg)$/i', $this->getOriginalImage()->getUrl())) {
             return 'image/jpg';
         }
 
-        if (preg_match('/\.(png)$/i', $this->getOriginalImage())) {
+        if (preg_match('/\.(png)$/i', $this->getOriginalImage()->getUrl())) {
             return 'image/png';
         }