+++ Plugin/CorrectImagesInAjaxResponse.php
@@ -2,95 +2,89 @@
 
 namespace Yireo\NextGenImages\Plugin;
 
-use Magento\Swatches\Helper\Data;
-use Yireo\NextGenImages\Browser\BrowserSupport;
+use Magento\Swatches\Helper\Data as SwatchesDataHelper;
 use Yireo\NextGenImages\Config\Config;
+use Yireo\NextGenImages\Image\ImageCollector;
 
 class CorrectImagesInAjaxResponse
 {
     /**
-     * @var BrowserSupport
-     */
-    private $browserSupport;
-
-    /**
      * @var Config
      */
     private $config;
-
+    
+    /**
+     * @var ImageCollector
+     */
+    private $imageCollector;
+    
     /**
      * CorrectImagesInAjaxResponse constructor.
      *
-     * @param BrowserSupport $browserSupport
      * @param Config $config
+     * @param ImageCollector $imageCollector
      */
     public function __construct(
-        BrowserSupport $browserSupport,
-        Config $config
+        Config $config,
+        ImageCollector $imageCollector
     ) {
-        $this->browserSupport = $browserSupport;
         $this->config = $config;
+        $this->imageCollector = $imageCollector;
     }
-
+    
     /**
-     * @param Data $dataHelper
+     * @param SwatchesDataHelper $dataHelper
      * @param array $data
      * @return mixed[]
      */
-    public function afterGetProductMediaGallery(Data $dataHelper, array $data): array
+    public function afterGetProductMediaGallery(SwatchesDataHelper $dataHelper, array $data): array
     {
         if (!$this->config->enabled()) {
             return $data;
         }
-
-        if (!$this->browserSupport->hasWebpSupport()) {
-            return $data;
-        }
-
+        
         return $this->replaceUrls($data);
     }
-
+    
     /**
-     * @param mixed[] $dataArray
+     * @param mixed[] $data
      * @return mixed[]
      */
-    private function replaceUrls(array $dataArray): array
+    private function replaceUrls(array $data): array
     {
-        if (empty($dataArray)) {
-            return $dataArray;
+        if (empty($data)) {
+            return $data;
         }
-
-        foreach ($dataArray as $name => $value) {
-            if (is_array($value)) {
-                $dataArray[$name] = $this->replaceUrls($value);
+        
+        $types = ['large', 'medium', 'small'];
+        foreach ($types as $type) {
+            if (empty($data[$type])) {
                 continue;
             }
-
-            if (!$this->isValidUrl($value)) {
-                continue;
+            
+            $images = $this->imageCollector->collect($data[$type]);
+            foreach ($images as $image) {
+                $data[$type . '_' . $image->getCode()] = $image->getUrl();
             }
-
-            // @todo: Shouldn't this be a nextgen image?
-            $dataArray[$name] = $value;
         }
-
-        return $dataArray;
-    }
-
-    /**
-     * @param string $url
-     * @return bool
-     */
-    private function isValidUrl(string $url): bool
-    {
-        if (!is_string($url)) {
-            return false;
+        
+        if (!isset($data['gallery'])) {
+            return $data;
         }
-
-        if (!preg_match('/\.(jpg|png)$/', $url)) {
-            return false;
+        
+        foreach ($data['gallery'] as $galleryIndex => $galleryImages) {
+            foreach ($types as $type) {
+                if (!isset($data[$type])) {
+                    continue;
+                }
+                
+                $images = $this->imageCollector->collect($data[$type]);
+                foreach ($images as $image) {
+                    $data['gallery'][$galleryIndex][$type . '_' . $image->getCode()] = $image->getUrl();
+                }
+            }
         }
-
-        return true;
+        
+        return $data;
     }
 }